Separate minidump creation into its own file
