+++ MWSE/MWSE.vcxproj
@@ -32,8 +32,10 @@
     <ClInclude Include="LuaStartGlobalScriptEvent.h" />
     <ClInclude Include="LuaUiSkillTooltipEvent.h" />
     <ClInclude Include="LuaUpdateBodyPartsForItemEvent.h" />
+    <ClInclude Include="Minidump.h" />
     <ClInclude Include="NIBoundingVolumeLua.h" />
     <ClInclude Include="NIBSAnimationNodeLua.h" />
+    <ClInclude Include="PatchNI.h" />
     <ClInclude Include="TES3AnimationGroupLua.h" />
     <ClInclude Include="TES3CrimeLua.h" />
     <ClInclude Include="TES3DialogueFilterContext.h" />
@@ -673,6 +675,7 @@
     <ClCompile Include="MGEPostShaders.cpp" />
     <ClCompile Include="MGEPostShadersLua.cpp" />
     <ClCompile Include="MGEUtilLua.cpp" />
+    <ClCompile Include="Minidump.cpp" />
     <ClCompile Include="MWSEConfig.cpp" />
     <ClCompile Include="MWSEUtilLua.cpp" />
     <ClCompile Include="NIBoundingVolumeLua.cpp" />
@@ -688,6 +691,7 @@
     <ClCompile Include="NISkinInstanceLua.cpp" />
     <ClCompile Include="NITextureEffectLua.cpp" />
     <ClCompile Include="NITextureLua.cpp" />
+    <ClCompile Include="PatchNI.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
