+++ MWSE/Minidump.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace mwse::minidump {
+	bool installHook();
+}
